refactor(turns): move shared turn/calendar logic behind boundaries
Move turn/calendar planning types and utilities that must run in both the app and Edge Function into
an Edge-compatible shared module boundary. The app can import that module through normal repo
conventions, and the Edge Function can import it through an explicit runtime-safe path or supported
import-map setup.

fix #118
